+++ druid/src/data.rs
@@ -19,6 +19,7 @@ use std::sync::Arc;
 
 use crate::kurbo::{self, ParamCurve};
 use crate::piet;
+use crate::shell::Scale;
 
 pub use druid_derive::Data;
 
@@ -244,6 +245,12 @@ impl<T0: Data, T1: Data, T2: Data, T3: Data, T4: Data, T5: Data> Data for (T0, T
     }
 }
 
+impl Data for Scale {
+    fn same(&self, other: &Self) -> bool {
+        self == other
+    }
+}
+
 impl Data for kurbo::Point {
     fn same(&self, other: &Self) -> bool {
         self.x.same(&other.x) && self.y.same(&other.y)
@@ -381,27 +388,16 @@ impl Data for piet::Color {
     }
 }
 
-//FIXME Vector::ptr_eq is not currently reliable; this is a temporary impl?
 #[cfg(feature = "im")]
 impl<T: Data> Data for im::Vector<T> {
     fn same(&self, other: &Self) -> bool {
-        // for reasons outlined in https://github.com/bodil/im-rs/issues/129,
-        // ptr_eq always returns false for small collections. This heuristic
-        // falls back to using equality for collections below some threshold.
-        // There may be a possibility of this returning false negatives, but
-        // not false positives; that's an acceptable outcome.
-
-        /* this is the impl I expected to use
-        const INLINE_LEN: usize = 48; // bytes available before first allocation;
-        let inline_capacity: usize = INLINE_LEN / std::mem::size_of::<T>();
-        if self.len() == other.len() && self.len() <= inline_capacity {
-            self.iter().zip(other.iter()).all(|(a, b)| a.same(b))
+        // if a vec is small enough that it doesn't require an allocation
+        // it is 'inline'; in this case a pointer comparison is meaningless.
+        if self.is_inline() {
+            self.len() == other.len() && self.iter().zip(other.iter()).all(|(a, b)| a.same(b))
         } else {
             self.ptr_eq(other)
         }
-        */
-
-        self.len() == other.len() && self.iter().zip(other.iter()).all(|(a, b)| a.same(b))
     }
 }