+++ druid/src/lib.rs
@@ -146,14 +146,14 @@ pub use piet::{Color, LinearGradient, RadialGradient, RenderContext, UnitPoint};
 pub use shell::{
     Application, Clipboard, ClipboardFormat, Cursor, Error as PlatformError, FileDialogOptions,
     FileInfo, FileSpec, FormatId, HotKey, KeyCode, KeyEvent, KeyModifiers, MouseButton,
-    MouseButtons, RawMods, SysMods, Text, TimerToken, WindowHandle,
+    MouseButtons, RawMods, Scale, SysMods, Text, TimerToken, WindowHandle,
 };
 
 pub use crate::core::WidgetPod;
 pub use app::{AppLauncher, WindowDesc};
 pub use app_delegate::{AppDelegate, DelegateCtx};
 pub use box_constraints::BoxConstraints;
-pub use command::{sys as commands, Command, Selector, Target};
+pub use command::{sys as commands, Command, Selector, SingleUse, Target};
 pub use contexts::{EventCtx, LayoutCtx, LifeCycleCtx, PaintCtx, Region, UpdateCtx};
 pub use data::Data;
 pub use env::{Env, Key, KeyOrValue, Value, ValueType};