+++ CHANGELOG.md
@@ -42,10 +42,15 @@ This means that druid no longer requires cairo on macOS and uses Core Graphics i
 - `MouseButtons` to `MouseEvent` to track which buttons are being held down during an event. ([#843] by [@xStrom])
 - `Env` and `Key` gained methods for inspecting an `Env` at runtime ([#880] by [@Zarenor])
 - `UpdateCtx::request_timer` and `UpdateCtx::request_anim_frame`. ([#898] by [@finnerale])
+- `LifeCycleCtx::request_timer`. ([#954] by [@xStrom])
+- `scale` method to `WinHandler`. ([#904] by [@xStrom])
+- `WinHandler::scale` method to inform of scale changes. ([#904] by [@xStrom])
 - `UpdateCtx::size` and `LifeCycleCtx::size`. ([#917] by [@jneem])
 - `WidgetExt::debug_widget_id`, for displaying widget ids on hover. ([#876] by [@cmyr])
 - `im` feature, with `Data` support for the [`im` crate](https://docs.rs/im/) collections. ([#924] by [@cmyr])
 - `im::Vector` support for the `List` widget. ([#940] by [@xStrom])
+- `LifeCycle::Size` event to inform widgets that their size changed. ([#953] by [@xStrom])
+- `Button::dynamic` constructor. ([#963] by [@totsteps])
 
 ### Changed
 
@@ -59,9 +64,12 @@ This means that druid no longer requires cairo on macOS and uses Core Graphics i
 - Global `Application` associated functions are instance methods instead, e.g. `Application::global().quit()` instead of the old `Application::quit()`. ([#763] by [@xStrom])
 - Timer events will only be delivered to the widgets that requested them. ([#831] by [@sjoshid])
 - `Event::Wheel` now contains a `MouseEvent` structure. ([#895] by [@teddemunnik])
+- The `WindowHandle::get_dpi` method got replaced by `WindowHandle::get_scale`. ([#904] by [@xStrom])
+- The `WinHandler::size` method now gets a `Size` in display points. ([#904] by [@xStrom])
 - `AppDelegate::command` now receives a `Target` instead of a `&Target`. ([#909] by [@xStrom])
 - `SHOW_WINDOW` and `CLOSE_WINDOW` commands now only use `Target` to determine the affected window. ([#928] by [@finnerale])
 - Replaced `NEW_WINDOW`, `SET_MENU` and `SHOW_CONTEXT_MENU` commands with methods on `EventCtx` and `DelegateCtx`. ([#931] by [@finnerale])
+- Replaced `Command::one_shot` and `::take_object` with a `SingleUse` payload wrapper type. ([#959] by [@finnerale])
 
 ### Deprecated
 
@@ -85,22 +93,27 @@ This means that druid no longer requires cairo on macOS and uses Core Graphics i
 - Keep hot state consistent with mouse position. ([#841] by [@xStrom])
 - Open file menu item works again. ([#851] by [@kindlychung])
 - Supply correct `LifeCycleCtx` to `Event::FocusChanged`. ([#878] by [@cmyr])
-- Windows: Termiate app when all windows have closed. ([#763] by [@xStrom])
+- Windows: Terminate app when all windows have closed. ([#763] by [@xStrom])
 - macOS: `Application::quit` now quits the run loop instead of killing the process. ([#763] by [@xStrom])
 - macOS/GTK/web: `MouseButton::X1` and `MouseButton::X2` clicks are now recognized. ([#843] by [@xStrom])
 - GTK: Support disabled menu items. ([#897] by [@jneem])
 - X11: Support individual window closing. ([#900] by [@xStrom])
 - X11: Support `Application::quit`. ([#900] by [@xStrom])
 - GTK: Support file filters in open/save dialogs. ([#903] by [@jneem])
+- GTK: Support DPI values other than 96. ([#904] by [@xStrom])
 - X11: Support key and mouse button state. ([#920] by [@jneem])
 - Routing `LifeCycle::FocusChanged` to descendant widgets. ([#925] by [@yrns])
 - Built-in open and save menu items now show the correct label and submit the right commands. ([#930] by [@finnerale])
+- Focus request handling is now predictable with the last request overriding earlier ones. ([#948] by [@xStrom])
+- Wheel events now properly update hot state. ([#951] by [@xStrom])
+- X11: Support mouse scrolling. ([#961] by [@jneem])
 
 ### Visual
 
 - Improved `Split` accuracy. ([#738] by [@xStrom])
 - Built-in widgets no longer stroke outside their `paint_rect`. ([#861] by [@jneem])
 - `Switch` toggles with animation when its data changes externally. ([#898] by [@finnerale])
+- Render progress bar correctly. ([#949] by [@scholtzan])
 
 ### Docs
 
@@ -112,6 +125,7 @@ This means that druid no longer requires cairo on macOS and uses Core Graphics i
 - Added new example for blocking functions. ([#840] by [@mastfissh])
 - Added a changelog containing development since the 0.5 release. ([#889] by [@finnerale])
 - Removed references to cairo on macOS. ([#943] by [@xStrom])
+- Updated screenshots in `README.md`. ([#967] by [@xStrom])
 
 ### Maintenance
 
@@ -122,6 +136,7 @@ This means that druid no longer requires cairo on macOS and uses Core Graphics i
 - GTK: Refactored `Application` to use the new structure. ([#892] by [@xStrom])
 - X11: Refactored `Application` to use the new structure. ([#894] by [@xStrom])
 - X11: Refactored `Window` to support some reentrancy and invalidation. ([#894] by [@xStrom])
+- Refactored DPI scaling. ([#904] by [@xStrom])
 - Added docs generation testing for all features. ([#942] by [@xStrom])
 
 ### Outside News
@@ -180,6 +195,7 @@ This means that druid no longer requires cairo on macOS and uses Core Graphics i
 [#898]: https://github.com/xi-editor/druid/pull/898
 [#900]: https://github.com/xi-editor/druid/pull/900
 [#903]: https://github.com/xi-editor/druid/pull/903
+[#904]: https://github.com/xi-editor/druid/pull/904
 [#905]: https://github.com/xi-editor/druid/pull/905
 [#909]: https://github.com/xi-editor/druid/pull/909
 [#917]: https://github.com/xi-editor/druid/pull/917
@@ -192,6 +208,15 @@ This means that druid no longer requires cairo on macOS and uses Core Graphics i
 [#940]: https://github.com/xi-editor/druid/pull/940
 [#942]: https://github.com/xi-editor/druid/pull/942
 [#943]: https://github.com/xi-editor/druid/pull/943
+[#948]: https://github.com/xi-editor/druid/pull/948
+[#949]: https://github.com/xi-editor/druid/pull/949
+[#951]: https://github.com/xi-editor/druid/pull/951
+[#953]: https://github.com/xi-editor/druid/pull/953
+[#954]: https://github.com/xi-editor/druid/pull/954
+[#959]: https://github.com/xi-editor/druid/pull/959
+[#961]: https://github.com/xi-editor/druid/pull/961
+[#963]: https://github.com/xi-editor/druid/pull/963
+[#967]: https://github.com/xi-editor/druid/pull/967
 
 ## [0.5.0] - 2020-04-01
 
@@ -223,6 +248,7 @@ Last release without a changelog :(
 [@Zarenor]: https://github.com/Zarenor
 [@yrns]: https://github.com/yrns
 [@jrmuizel]: https://github.com/jrmuizel
+[@scholtzan]: https://github.com/scholtzan
 
 [Unreleased]: https://github.com/xi-editor/druid/compare/v0.5.0...master
 [0.5.0]: https://github.com/xi-editor/druid/compare/v0.4.0...v0.5.0